+++ R/filter-repeat-visits.r
@@ -17,11 +17,12 @@
 #' @param min_obs integer; minimum number of observations required for each
 #'   site.
 #' @param max_obs integer; maximum number of observations allowed for each site.
+#' @param annual_closure logical; whether the entire year should be treated as
+#'   the period of closure (the default). This can be useful, for example, if 
+#'   the data have been subset to a period of closure prior to calling 
+#'   [filter_repeat_visits()].
 #' @param n_days integer; number of days defining the temporal length of
 #'   closure. Ignored if `annual_closure = TRUE`.
-#' @param annual_closure logical; whether the entire year should be treated as
-#'   the period of closure. This can be useful, for example, if data are from
-#'   one season (e.g. breeding) across multiple years.
 #' @param date_var character; column name of the variable in `x` containing the
 #'   date. This column should either be in `Date` format or convertible to
 #'   `Date` format with [as.Date()].
@@ -71,7 +72,7 @@
 #'                        collapse = TRUE)
 #' filter_repeat_visits(ebd_zf, n_days = 30)
 filter_repeat_visits <- function(x, min_obs = 2L, max_obs = 10L, 
-                                 n_days = 14L, annual_closure = FALSE,
+                                 annual_closure = TRUE, n_days = NULL,
                                  date_var = "observation_date",
                                  site_vars = c("locality_id", "observer_id"),
                                  ll_digits = 6L) {
@@ -81,14 +82,26 @@ filter_repeat_visits <- function(x, min_obs = 2L, max_obs = 10L,
   }
   stopifnot(is.data.frame(x))
   stopifnot(is_integer(min_obs), length(min_obs) == 1, isTRUE(min_obs > 0),
-            is_integer(max_obs), length(max_obs) == 1, isTRUE(max_obs > 0),
-            is_integer(n_days), length(n_days) == 1, isTRUE(n_days > 0)) 
+            is_integer(max_obs), length(max_obs) == 1, isTRUE(max_obs > 0)) 
   stopifnot(is.logical(annual_closure), length(annual_closure) == 1)
   stopifnot(is.character(date_var), length(date_var) == 1,
             date_var %in% names(x))
   stopifnot(is.character(site_vars), all(site_vars %in% names(x)))
   stopifnot(is_integer(ll_digits), length(ll_digits) == 1,
             isTRUE(ll_digits > 0))
+  # must define period of closure if annual_closure = FALE
+  if (isTRUE(annual_closure)) {
+    if (!is.null(n_days)) {
+      warning(sprintf("n_days = %i but will be ignored %s", n_days,
+                      "because annual_closure is TRUE"))
+    }
+  } else {
+    if (is.null(n_days)) {
+      stop(paste("When annual_closure is FALSE, n_days must be used to specify",
+                 "the length of the period of closure."))
+    }
+    stopifnot(is_integer(n_days), length(n_days) == 1, isTRUE(n_days > 0))
+  }
   # can't have variables overlapping with added variables
   prohibit <- c("site", "closure_id", "n_observations")
   if (any(prohibit %in% names(x))) {