+++ NEWS.md
@@ -5,6 +5,8 @@
 - Better handling of file paths with `prefix` argument in `auk_split()`
 - Fixed bug causing undescribed species to be dropped by `auk_rollup()`
 - Add a `ll_digits` argument to `filter_repeat_visits()` to round lat/lng prior to identifying sites
+- Change of default parameters to `filter_repeat_visits()`
+- `auk_bbox()` now takes sf/raster spatial objects and grabs bbox from them
 
 # auk 0.4.0